refactor(scraping): Extract connection actions
