+++ tests/test_scraping.py
@@ -1024,831 +1024,6 @@ class TestDetectConnectionState:
         )
 
 
-class TestConnectWithPerson:
-    def _mock_scrape(
-        self, profile_text: str, *, follow_up_text: str | None = None
-    ) -> AsyncMock:
-        """Return a mock for scrape_person.
-
-        When ``follow_up_text`` is given, the second call returns that text
-        — used to simulate verification re-reads after an action.
-        """
-        first = {
-            "url": "https://www.linkedin.com/in/testuser/",
-            "sections": {"main_profile": profile_text},
-        }
-        if follow_up_text is None:
-            return AsyncMock(return_value=first)
-        second = {
-            "url": "https://www.linkedin.com/in/testuser/",
-            "sections": {"main_profile": follow_up_text},
-        }
-        return AsyncMock(side_effect=[first, second])
-
-    @staticmethod
-    def _signals(
-        invite: bool = False,
-        compose: bool = False,
-        edit: bool = False,
-        labeled_action: bool = False,
-        labeled_anchor: bool = False,
-        incoming_row: bool = False,
-    ) -> ActionSignals:
-        return ActionSignals(
-            has_invite_anchor=invite,
-            has_compose_anchor_in_action_root=compose,
-            has_edit_intro_anchor=edit,
-            has_labeled_action_button=labeled_action,
-            has_labeled_action_anchor=labeled_anchor,
-            has_incoming_action_row=incoming_row,
-        )
-
-    async def test_connectable_navigates_deeplink_and_verifies(self, mock_page):
-        """Connect via deeplink: dialog opens, submit succeeds, anchor disappears."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Jane\n\n· 3rd\n\nEngineer\n\nConnect\nMore\nAbout\n"
-        post_text = "Jane\n\n· 3rd\n\nEngineer\n\nMessage\nPending\nMore\nAbout\n"
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                self._mock_scrape(text, follow_up_text=post_text),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=[self._signals(invite=True), self._signals()],
-            ),
-            patch.object(
-                PageNavigator,
-                "_navigate_to_page",
-                new_callable=AsyncMock,
-            ) as mock_nav,
-            patch.object(
-                extractor,
-                "_dialog_is_open",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch.object(
-                extractor,
-                "_click_dialog_primary_button",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connected"
-        mock_nav.assert_awaited_once()
-        await_args = mock_nav.await_args
-        assert await_args is not None
-        assert "preload/custom-invite" in await_args.args[0]
-
-    async def test_connectable_send_failed_when_anchor_persists(self, mock_page):
-        """Dialog submitted but profile still exposes Connect → send_failed."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Jane\n\n· 3rd\n\nEngineer\n\nConnect\nMore\nAbout\n"
-
-        with (
-            patch.object(
-                extractor, "scrape_person", self._mock_scrape(text, follow_up_text=text)
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=[self._signals(invite=True), self._signals(invite=True)],
-            ),
-            patch.object(PageNavigator, "_navigate_to_page", new_callable=AsyncMock),
-            patch.object(
-                extractor, "_dialog_is_open", new_callable=AsyncMock, return_value=True
-            ),
-            patch.object(
-                extractor,
-                "_click_dialog_primary_button",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "send_failed"
-
-    async def test_premium_upsell_message_reads_linkedin_dialog_text(self, mock_page):
-        """Premium upsell detection returns LinkedIn's raw dialog text."""
-        extractor = LinkedInExtractor(mock_page)
-        premium_link = MagicMock()
-        premium_link.wait_for = AsyncMock(return_value=None)
-        premium_link.is_visible = AsyncMock(return_value=True)
-        premium_link.inner_text = AsyncMock(return_value="fallback")
-        premium_link.first = premium_link
-        mock_page.locator.return_value = premium_link
-        mock_page.evaluate = AsyncMock(
-            return_value="Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium"
-        )
-
-        result = await extractor._get_premium_upsell_message(timeout=1234)
-
-        assert (
-            result
-            == "Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium"
-        )
-        mock_page.locator.assert_called_once_with(
-            'dialog[open] a[href*="/premium/"], [role="dialog"] a[href*="/premium/"]'
-        )
-        premium_link.wait_for.assert_awaited_once_with(state="visible", timeout=1234)
-
-    async def test_submit_invite_dialog_reports_premium_after_add_note(self, mock_page):
-        """Add-note Premium upsell is a note-limit block, not no-dialog."""
-        from patchright.async_api import TimeoutError as PlaywrightTimeoutError
-
-        extractor = LinkedInExtractor(mock_page)
-        textarea = MagicMock()
-        textarea.count = AsyncMock(return_value=0)
-        add_note_button = MagicMock()
-        add_note_button.click = AsyncMock(return_value=None)
-        buttons = MagicMock()
-        buttons.count = AsyncMock(return_value=3)
-        buttons.nth.return_value = add_note_button
-
-        def locator_for(selector: str):
-            return textarea if "textarea" in selector else buttons
-
-        mock_page.locator.side_effect = locator_for
-        mock_page.wait_for_selector = AsyncMock(
-            side_effect=PlaywrightTimeoutError("textarea timeout")
-        )
-
-        with (
-            patch.object(
-                extractor, "_dialog_is_open", new_callable=AsyncMock, return_value=True
-            ),
-            patch.object(
-                extractor,
-                "_get_premium_upsell_message",
-                new_callable=AsyncMock,
-                return_value="Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium",
-            ) as mock_message,
-            patch.object(
-                extractor, "_dismiss_dialog", new_callable=AsyncMock
-            ) as mock_dismiss,
-        ):
-            result = await extractor._submit_invite_dialog("Hello")
-
-        assert result == (
-            False,
-            False,
-            "Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium",
-        )
-        add_note_button.click.assert_awaited_once()
-        mock_message.assert_awaited_once()
-        mock_dismiss.assert_awaited_once()
-
-    async def test_submit_invite_dialog_reports_premium_after_send_click_failure(
-        self, mock_page
-    ):
-        """Premium upsell intercepting the Send click is a note-limit block.
-
-        When LinkedIn swaps the invite dialog for the Premium upsell at the
-        moment of submit, the original primary button is detached or pointer-
-        event covered, so ``_click_dialog_primary_button`` and the keyboard
-        fallback both fail. Without the post-click upsell probe the caller
-        would dismiss the dialog and report ``connect_unavailable`` even
-        though LinkedIn's raw quota message is sitting in the visible modal.
-        """
-        extractor = LinkedInExtractor(mock_page)
-
-        # Textarea already exposed so the reveal/fill branch succeeds and the
-        # test focuses on the post-submit failure path.
-        textarea = MagicMock()
-        textarea.count = AsyncMock(return_value=1)
-        textarea.first = textarea
-        textarea.fill = AsyncMock()
-
-        buttons = MagicMock()
-        buttons.count = AsyncMock(return_value=2)
-        primary_button = MagicMock()
-        primary_button.focus = AsyncMock()
-        buttons.nth.return_value = primary_button
-
-        def locator_for(selector: str):
-            return textarea if "textarea" in selector else buttons
-
-        mock_page.locator.side_effect = locator_for
-        mock_page.keyboard = MagicMock()
-        mock_page.keyboard.press = AsyncMock()
-
-        message = "You're out of free custom notes. Bypass the limit with Premium..."
-
-        with (
-            patch.object(
-                extractor,
-                "_dialog_is_open",
-                new_callable=AsyncMock,
-                # First call: dialog open at entry. Second call: still open
-                # after the keyboard fallback, so sent remains False.
-                side_effect=[True, True],
-            ),
-            patch.object(
-                extractor,
-                "_fill_dialog_textarea",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch.object(
-                extractor,
-                "_click_dialog_primary_button",
-                new_callable=AsyncMock,
-                return_value=False,
-            ),
-            patch.object(
-                extractor,
-                "_get_premium_upsell_message",
-                new_callable=AsyncMock,
-                return_value=message,
-            ) as mock_message,
-            patch.object(
-                extractor, "_dismiss_dialog", new_callable=AsyncMock
-            ) as mock_dismiss,
-        ):
-            result = await extractor._submit_invite_dialog("Hello")
-
-        assert result == (False, False, message)
-        mock_message.assert_awaited_once()
-        mock_dismiss.assert_awaited_once()
-
-    async def test_connectable_no_dialog_returns_connect_unavailable(self, mock_page):
-        """Deeplink opened but no dialog appeared → connect_unavailable."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Jane\n\n· 3rd\n\nEngineer\n\nConnect\nMore\nAbout\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(invite=True),
-            ),
-            patch.object(PageNavigator, "_navigate_to_page", new_callable=AsyncMock),
-            patch.object(
-                extractor, "_dialog_is_open", new_callable=AsyncMock, return_value=False
-            ),
-            patch.object(extractor, "_dismiss_dialog", new_callable=AsyncMock),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connect_unavailable"
-
-    async def test_returns_already_connected_via_anchor(self, mock_page):
-        """1st-degree detected via /messaging/compose anchor."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Collin\n\n· 1st\n\nEngineer\n\nMessage\nMore\nAbout\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(compose=True),
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "already_connected"
-
-    async def test_returns_self_profile_via_edit_intro_anchor(self, mock_page):
-        """Editing-your-own-profile anchor blocks connect attempts."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Daniel\n\nEdit profile\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(edit=True),
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connect_unavailable"
-        assert "own profile" in result["message"]
-
-    async def test_connect_via_more_menu(self, mock_page):
-        """Follow-primary profile with Connect under More: detection sees
-        no invite anchor initially, _open_more_menu surfaces it, deeplink
-        fires."""
-        extractor = LinkedInExtractor(mock_page)
-        # Pre-More: Follow primary, Connect hidden under the More dropdown.
-        pre = "Christian\n\n· 2nd\n\nFounder\n\nFollow\nMessage\nMore\n"
-        post = "Christian\n\n· 2nd\n\nFounder\n\nMessage\nPending\nMore\n"
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                self._mock_scrape(pre, follow_up_text=post),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                # 1st: follow_only (compose+labeled, no invite).
-                # 2nd: post-More reread reveals invite anchor.
-                # 3rd: post-deeplink verification — invite anchor gone.
-                side_effect=[
-                    self._signals(compose=True, labeled_action=True),
-                    self._signals(invite=True, compose=True, labeled_action=True),
-                    self._signals(),
-                ],
-            ),
-            patch.object(
-                extractor,
-                "_open_more_menu",
-                new_callable=AsyncMock,
-                return_value=True,
-            ) as mock_open_more,
-            patch.object(
-                PageNavigator, "_navigate_to_page", new_callable=AsyncMock
-            ) as mock_nav,
-            patch.object(
-                extractor,
-                "_dialog_is_open",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch.object(
-                extractor,
-                "_click_dialog_primary_button",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connected"
-        mock_open_more.assert_awaited_once()
-        # Deeplink fired exactly once.
-        assert mock_nav.await_count == 1
-        await_args = mock_nav.await_args
-        assert await_args is not None
-        assert "preload/custom-invite" in await_args.args[0]
-
-    async def test_follow_only_after_more_does_not_send(self, mock_page):
-        """Pending or genuinely follow-only profile: invite anchor never
-        appears even after More-menu open. Critical write-gate guardrail —
-        no deeplink fires, no connection request goes out."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Public Figure\n\n· 3rd+\n\nCEO\n\nFollow\nMessage\nMore\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                # Both reads (initial + post-More) show no invite anchor.
-                side_effect=[
-                    self._signals(compose=True, labeled_action=True),
-                    self._signals(compose=True, labeled_action=True),
-                ],
-            ),
-            patch.object(
-                extractor,
-                "_open_more_menu",
-                new_callable=AsyncMock,
-                return_value=True,
-            ) as mock_open_more,
-            patch.object(
-                PageNavigator, "_navigate_to_page", new_callable=AsyncMock
-            ) as mock_nav,
-            patch.object(
-                extractor, "_submit_invite_dialog", new_callable=AsyncMock
-            ) as mock_submit,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connect_unavailable"
-        assert result.get("note_sent") is False or "note_sent" not in result
-        mock_open_more.assert_awaited_once()
-        # Critical: deeplink must NOT fire and dialog must NOT be submitted.
-        mock_nav.assert_not_awaited()
-        mock_submit.assert_not_awaited()
-
-    async def test_follow_only_with_note_reports_note_limit_from_deeplink_probe(
-        self, mock_page
-    ):
-        """A requested note may reveal Premium quota without submitting."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Public Figure\n\n· 3rd+\n\nCEO\n\nFollow\nMessage\nMore\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=[
-                    self._signals(compose=True, labeled_action=True),
-                    self._signals(compose=True, labeled_action=True),
-                ],
-            ),
-            patch.object(
-                extractor,
-                "_open_more_menu",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch.object(
-                PageNavigator, "_navigate_to_page", new_callable=AsyncMock
-            ) as mock_nav,
-            patch.object(
-                extractor,
-                "_probe_invite_note_limit",
-                new_callable=AsyncMock,
-                return_value="Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium",
-            ) as mock_probe,
-            patch.object(
-                extractor, "_submit_invite_dialog", new_callable=AsyncMock
-            ) as mock_submit,
-        ):
-            result = await extractor.connect_with_person("testuser", note="Hello")
-
-        assert result["status"] == "custom_note_limit_reached"
-        assert (
-            result["message"]
-            == "Wysyłaj nieograniczoną liczbę spersonalizowanych zaproszeń dzięki Premium"
-        )
-        assert result["note_sent"] is False
-        mock_nav.assert_awaited_once()
-        mock_probe.assert_awaited_once()
-        mock_submit.assert_not_awaited()
-
-    async def test_more_menu_unavailable_does_not_send(self, mock_page):
-        """Action root present but no More button (unusual but possible):
-        _open_more_menu returns False, no retry, no deeplink fires."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Public Figure\n\n· 3rd+\n\nCEO\n\nFollow\nMessage\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(compose=True, labeled_action=True),
-            ),
-            patch.object(
-                extractor,
-                "_open_more_menu",
-                new_callable=AsyncMock,
-                return_value=False,
-            ),
-            patch.object(
-                PageNavigator, "_navigate_to_page", new_callable=AsyncMock
-            ) as mock_nav,
-            patch.object(
-                extractor, "_submit_invite_dialog", new_callable=AsyncMock
-            ) as mock_submit,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "connect_unavailable"
-        mock_nav.assert_not_awaited()
-        mock_submit.assert_not_awaited()
-
-    async def test_returns_pending(self, mock_page):
-        """Profile with a pending invitation: detected via labeled <a> in
-        the action root. Returns status='pending' without firing the
-        deeplink (LinkedIn would only show 'already invited' anyway)."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Frank\n\n· 3rd\n\nFounder\n\nMessage\nPending\nMore\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(compose=True, labeled_anchor=True),
-            ),
-            patch.object(
-                PageNavigator, "_navigate_to_page", new_callable=AsyncMock
-            ) as mock_nav,
-            patch.object(
-                extractor, "_submit_invite_dialog", new_callable=AsyncMock
-            ) as mock_submit,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "pending"
-        # No write-path side effects.
-        mock_nav.assert_not_awaited()
-        mock_submit.assert_not_awaited()
-
-    async def test_returns_incoming_request_accepted(self, mock_page):
-        """Structural detection + structural accept click, German locale."""
-        extractor = LinkedInExtractor(mock_page)
-        pre = "Eric\n\n· 2.\n\nAachen\n\nAnnehmen\nIgnorieren\nMehr\nInfo\n"
-        post = "Eric\n\n· 1.\n\nAachen\n\nNachricht\nMehr\nInfo\n"
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                self._mock_scrape(pre, follow_up_text=post),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=[
-                    self._signals(incoming_row=True),
-                    self._signals(compose=True),
-                ],
-            ),
-            patch.object(
-                extractor,
-                "_click_incoming_accept",
-                new_callable=AsyncMock,
-                return_value=True,
-            ) as mock_accept,
-            patch.object(
-                PageNavigator,
-                "_navigate_to_page",
-                new_callable=AsyncMock,
-            ) as mock_nav,
-            patch.object(
-                extractor,
-                "_submit_invite_dialog",
-                new_callable=AsyncMock,
-            ) as mock_submit,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "accepted"
-        mock_accept.assert_awaited_once()
-        mock_nav.assert_not_awaited()
-        mock_submit.assert_not_awaited()
-
-    async def test_incoming_request_send_failed_when_click_fails(self, mock_page):
-        """Structural accept click did not land; no locale-text guessing —
-        report send_failed without navigating or clicking by text."""
-        extractor = LinkedInExtractor(mock_page)
-        pre = "Eric\n\n· 2.\n\nAachen\n\nAnnehmen\nIgnorieren\nMehr\nInfo\n"
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                self._mock_scrape(pre),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(incoming_row=True),
-            ),
-            patch.object(
-                extractor,
-                "_click_incoming_accept",
-                new_callable=AsyncMock,
-                return_value=False,
-            ),
-            patch.object(
-                extractor,
-                "click_button_by_text",
-                new_callable=AsyncMock,
-                return_value=True,
-            ) as mock_text_click,
-            patch.object(
-                PageNavigator,
-                "_navigate_to_page",
-                new_callable=AsyncMock,
-            ) as mock_nav,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "send_failed"
-        mock_nav.assert_not_awaited()
-        # No text-based clicking on the destructive accept path.
-        mock_text_click.assert_not_awaited()
-
-    async def test_incoming_request_send_failed_when_no_first_degree(self, mock_page):
-        """Accept clicked but profile never transitions to 1st-degree."""
-        extractor = LinkedInExtractor(mock_page)
-        pre = "Eric\n\n· 2.\n\nAachen\n\nAnnehmen\nIgnorieren\nMehr\nInfo\n"
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                AsyncMock(
-                    return_value={
-                        "url": "https://www.linkedin.com/in/testuser/",
-                        "sections": {"main_profile": pre},
-                    }
-                ),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(incoming_row=True),
-            ),
-            patch.object(
-                extractor,
-                "_click_incoming_accept",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch(
-                "linkedin_mcp_server.scraping.extractor.asyncio.sleep",
-                new_callable=AsyncMock,
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "send_failed"
-
-    async def test_incoming_request_accepted_on_settle_retry(self, mock_page):
-        """The first post-click read still renders the old top card;
-        the settle retry sees the 1st-degree state and reports accepted."""
-        extractor = LinkedInExtractor(mock_page)
-        pre = "Eric\n\n· 2.\n\nAachen\n\nAnnehmen\nIgnorieren\nMehr\nInfo\n"
-        post = "Eric\n\n· 1.\n\nAachen\n\nNachricht\nMehr\nInfo\n"
-        page = {
-            "url": "https://www.linkedin.com/in/testuser/",
-            "sections": {"main_profile": pre},
-        }
-        page_post = {
-            "url": "https://www.linkedin.com/in/testuser/",
-            "sections": {"main_profile": post},
-        }
-
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                AsyncMock(side_effect=[page, page, page_post]),
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=[
-                    self._signals(incoming_row=True),
-                    self._signals(incoming_row=True),
-                    self._signals(compose=True),
-                ],
-            ),
-            patch.object(
-                extractor,
-                "_click_incoming_accept",
-                new_callable=AsyncMock,
-                return_value=True,
-            ),
-            patch(
-                "linkedin_mcp_server.scraping.extractor.asyncio.sleep",
-                new_callable=AsyncMock,
-            ) as mock_sleep,
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "accepted"
-        mock_sleep.assert_awaited_once()
-
-    async def test_returns_unavailable_when_no_signals_and_text(self, mock_page):
-        """No structural signals, no actionable text → connect_unavailable."""
-        extractor = LinkedInExtractor(mock_page)
-        text = "Public Figure\n\n· 3rd+\n\nCEO\n\nFollow\nMore\nAbout\n"
-
-        with (
-            patch.object(extractor, "scrape_person", self._mock_scrape(text)),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                return_value=self._signals(),
-            ),
-            patch.object(PageNavigator, "_navigate_to_page", new_callable=AsyncMock),
-            patch.object(
-                extractor, "_dialog_is_open", new_callable=AsyncMock, return_value=False
-            ),
-            patch.object(extractor, "_dismiss_dialog", new_callable=AsyncMock),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        # follow_only path goes through deeplink; no dialog opens → unavailable
-        assert result["status"] == "connect_unavailable"
-
-    async def test_returns_unavailable_on_empty_page(self, mock_page):
-        extractor = LinkedInExtractor(mock_page)
-
-        with patch.object(
-            extractor,
-            "scrape_person",
-            AsyncMock(
-                return_value={
-                    "url": "https://www.linkedin.com/in/testuser/",
-                    "sections": {},
-                }
-            ),
-        ):
-            result = await extractor.connect_with_person("testuser")
-
-        assert result["status"] == "unavailable"
-
-    async def test_submit_invite_dialog_handles_two_button_gating_dialog(
-        self, mock_page
-    ):
-        """Two-button "Add a note to your invitation?" gating dialog (issue
-        #455): nth(0) is "Add a note", nth(1) is "Send without a note".
-
-        Asserts the secondary-button click that reveals the textarea fires
-        even with btn_count == 2 (legacy guard required >= 3 and skipped
-        the click, leaving the textarea unmounted)."""
-        extractor = LinkedInExtractor(mock_page)
-
-        # Track each button click so we can assert the "Add a note" path
-        # was taken to reveal the textarea.
-        clicks: list[int] = []
-
-        textarea_visible = {"value": False}
-
-        # Two button locators inside the gating dialog: nth(0) "Add a
-        # note" reveals the textarea, nth(1) "Send without a note".
-        button_locators = [MagicMock(), MagicMock()]
-        for idx, btn in enumerate(button_locators):
-
-            def make_click(i: int):
-                async def _click(*args, **kwargs):
-                    clicks.append(i)
-                    if i == 0:
-                        textarea_visible["value"] = True
-                    return None
-
-                return _click
-
-            btn.click = AsyncMock(side_effect=make_click(idx))
-            btn.focus = AsyncMock()
-
-        button_collection = MagicMock()
-        button_collection.count = AsyncMock(return_value=2)
-        button_collection.nth = MagicMock(side_effect=lambda i: button_locators[i])
-
-        textarea_locator = MagicMock()
-        textarea_locator.count = AsyncMock(
-            side_effect=lambda: 1 if textarea_visible["value"] else 0
-        )
-        textarea_locator.first = textarea_locator
-        textarea_locator.fill = AsyncMock()
-
-        # Route page.locator() calls by selector — buttons vs textarea —
-        # so the gating dialog's button collection is distinguishable
-        # from the textarea probe.
-        def locator_router(selector: str):
-            if "textarea" in selector:
-                return textarea_locator
-            return button_collection
-
-        mock_page.locator = MagicMock(side_effect=locator_router)
-        mock_page.wait_for_selector = AsyncMock()
-        mock_page.keyboard = MagicMock()
-        mock_page.keyboard.press = AsyncMock()
-
-        with (
-            patch.object(
-                extractor, "_dialog_is_open", new_callable=AsyncMock, return_value=True
-            ),
-            patch.object(
-                extractor,
-                "_get_premium_upsell_message",
-                new_callable=AsyncMock,
-                return_value=None,
-            ),
-        ):
-            (
-                submitted,
-                note_sent,
-                note_limit_message,
-            ) = await extractor._submit_invite_dialog("Hi from a test")
-
-        assert submitted is True
-        assert note_sent is True
-        assert note_limit_message is None
-        # Clicked "Add a note" (index 0) to reveal the textarea, then the
-        # primary button (index 1) to send.
-        assert clicks == [0, 1]
-        textarea_locator.fill.assert_awaited_once()
-
-
 class TestScrapeCompany:
     async def test_a_pasted_company_link_reaches_the_canonical_company_url(
         self, mock_page
@@ -6958,18 +6133,6 @@ class TestMessageConfirmation:
         )
 
 
-def _no_signals() -> ActionSignals:
-    """Every structural signal absent, which is all these tests need."""
-    return ActionSignals(
-        has_invite_anchor=False,
-        has_compose_anchor_in_action_root=False,
-        has_edit_intro_anchor=False,
-        has_labeled_action_button=False,
-        has_labeled_action_anchor=False,
-        has_incoming_action_row=False,
-    )
-
-
 class TestEveryNormalizedEntryPoint:
     """Each method that was rewired, refusing a value that redirects the path.
 
@@ -6990,39 +6153,6 @@ class TestEveryNormalizedEntryPoint:
             patch.object(SectionCapture, "extract_page", new_callable=AsyncMock),
             patch.object(PageNavigator, "_navigate_to_page", new_callable=AsyncMock),
         )
-
-    async def test_connect_with_person_normalizes_before_its_own_downstream_use(
-        self, mock_page
-    ):
-        """The traversal case cannot see this one.
-
-        scrape_person normalizes too, so removing connect_with_person's own call
-        still raises on "../../feed". A full URL is what separates them: the
-        scrape would succeed while the invite deeplink and the action-signal
-        selectors kept receiving the URL where they expect the vanity.
-        """
-        extractor = LinkedInExtractor(mock_page)
-        seen: list[str] = []
-        with (
-            patch.object(
-                extractor,
-                "scrape_person",
-                new_callable=AsyncMock,
-                return_value={"sections": {"main_profile": "text"}},
-            ),
-            patch.object(
-                extractor,
-                "_read_action_signals",
-                new_callable=AsyncMock,
-                side_effect=lambda username: seen.append(username) or _no_signals(),
-            ),
-            patch.object(PageNavigator, "_navigate_to_page", new_callable=AsyncMock),
-        ):
-            await extractor.connect_with_person(
-                "https://de.linkedin.com/in/williamhgates"
-            )
-
-        assert seen == ["williamhgates"]
 
     @pytest.mark.parametrize(
         "method,args,kwargs",
